+++ packages/core/styles/src/packages/atomic/popover/index.ts
@@ -5,3 +5,5 @@ export const popover = {
   classnames: { ...styles },
   ...popoverSprinkle,
 };
+
+export type { PopoverSprinkle } from "./popover.sprinkle.types";