+++ packages/react/src/atomic/Box/src/Box.tsx
@@ -20,6 +20,8 @@ const Box = forwardRef(
       boxSizing,
     });
 
+    console.log("color", rest.color);
+
     return <As ref={ref} className={className} style={style} {...otherProps} />;
   }
 ) as PolymorphicForwardRefComponent<"div", BoxProps>;