+++ packages/react/src/atomic/Box/src/box.types.ts
@@ -1,7 +1,7 @@
 import { ReactNode, HTMLAttributes } from "react";
 import { BoxSprinkle } from "@nimbus-ds/styles";
 
-type BoxExtends = BoxSprinkle & HTMLAttributes<HTMLElement>;
+type BoxExtends = BoxSprinkle & Omit<HTMLAttributes<HTMLElement>, "color">;
 
 export interface BoxProps extends BoxExtends {
   /** Element to be rendered inside the Box component */